Fixed as per comments
