+++ assignment-impl/src/main/java/com/epam/rd/davydova/assignment/config/ComponentConfig.java
@@ -1,6 +1,9 @@
 package com.epam.rd.davydova.assignment.config;
 
-import org.springframework.context.annotation.*;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Profile;
 
 import javax.persistence.EntityManager;
 import javax.persistence.Persistence;
@@ -9,7 +12,8 @@ import javax.persistence.Persistence;
  * This is class for configuration of components that use EntityManager
  */
 @Configuration
-@ComponentScan("com.epam.rd.davydova.assignment.domain")
+@Profile("!local")
+@ComponentScan("com.epam.rd.davydova.assignment")
 public class ComponentConfig {
 
     @Bean
