+++ assignment-impl/src/main/java/com/epam/rd/davydova/assignment/config/MessageSourceConfig.java
@@ -1,14 +1,16 @@
 package com.epam.rd.davydova.assignment.config;
 
 import org.springframework.context.MessageSource;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.*;
 import org.springframework.context.support.ReloadableResourceBundleMessageSource;
 
 /**
  * This is class for configuration of internationalization
  */
 @Configuration
+@Profile("local")
+@ComponentScan("com.epam.rd.davydova.assignment.service.stub " +
+        "&& com.epam.rd.davydova.assignment.domain.stub")
 public class MessageSourceConfig {
 
     @Bean
