--- assignment-impl/src/main/java/com/epam/rd/davydova/assignment/domain/service/CustomerService.java
@@ -1,179 +1,0 @@
-package com.epam.rd.davydova.assignment.domain.service;
-
-import com.epam.rd.davydova.assignment.domain.model.Customer;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
-import javax.persistence.EntityManager;
-import javax.persistence.EntityTransaction;
-import java.util.List;
-import java.util.Optional;
-
-/**
- * This is a class for operations with Customer entity and database
- */
-@Slf4j
-@Data
-@NoArgsConstructor
-@Service
-public class CustomerService {
-    private EntityManager entityManager;
-
-    @Autowired
-    public CustomerService(EntityManager entityManager) {
-        this.entityManager = entityManager;
-    }
-
-    /**
-     * Add customer to database
-     *
-     * @param customerName customer instance
-     */
-    public void add(String customerName, String phone) {
-        EntityTransaction transaction = null;
-        try {
-            transaction = entityManager.getTransaction();
-            transaction.begin();
-            var customer = new Customer();
-            customer.setCustomerName(customerName);
-            customer.setPhone(phone);
-            entityManager.persist(customer);
-            transaction.commit();
-        } catch (Exception e) {
-            if (transaction != null) {
-                transaction.rollback();
-            }
-            log.error("Customer is not added. Exception is: ", e);
-        }
-    }
-
-    /**
-     * Find customer by their name
-     *
-     * @param customerName customer's name
-     * @return customer instance
-     */
-    public Optional<Customer> findBy(String customerName) {
-        EntityTransaction transaction = null;
-        try {
-            transaction = entityManager.getTransaction();
-            transaction.begin();
-            var customer = (Customer) entityManager
-                    .createNamedQuery(Customer.FIND_CUSTOMER_BY_NAME)
-                    .setParameter(1, customerName)
-                    .getSingleResult();
-            transaction.commit();
-            var foundCustomer = Optional.ofNullable(customer);
-            if (foundCustomer.isPresent()) {
-                log.info("Customer is found");
-                return foundCustomer;
-            }
-        } catch (Exception e) {
-            if (transaction != null) {
-                transaction.rollback();
-            }
-            log.error("Customer is not found. Exception is: ", e);
-        }
-        return Optional.empty();
-    }
-
-    /**
-     * Find customer by Id
-     *
-     * @param customerId customer Id
-     * @return Optional of customer instance
-     */
-    public Optional<Customer> findBy(int customerId) {
-        EntityTransaction transaction = null;
-        try {
-            transaction = entityManager.getTransaction();
-            transaction.begin();
-            var customer = entityManager.find(Customer.class, customerId);
-            transaction.commit();
-            var foundCustomer = Optional.ofNullable(customer);
-            if (foundCustomer.isPresent()) {
-                log.info("Customer is found");
-                return foundCustomer;
-            }
-        } catch (Exception e) {
-            if (transaction != null) {
-                transaction.rollback();
-            }
-            log.error("Customer is not found. Exception is: ", e);
-        }
-        return Optional.empty();
-    }
-
-    /**
-     * Find all customers
-     *
-     * @return Optional of List of customers
-     */
-    public Optional<List> findAll() {
-        EntityTransaction transaction = null;
-        try {
-            transaction = entityManager.getTransaction();
-            transaction.begin();
-            var customerList = entityManager.createNamedQuery(Customer.FIND_ALL_CUSTOMERS).getResultList();
-            transaction.commit();
-            var foundCustomerList = Optional.ofNullable(customerList);
-            if (foundCustomerList.isPresent()) {
-                log.info("List of customers is found");
-                return foundCustomerList;
-            }
-        } catch (Exception e) {
-            if (transaction != null) {
-                transaction.rollback();
-            }
-            log.error("List of customers is not found. Exception is: ", e);
-        }
-        return Optional.empty();
-    }
-
-    /**
-     * Update customer's name and phone number
-     *
-     * @param customerId customer Id
-     * @param phone      phone number
-     */
-    public void update(int customerId, String phone) {
-        EntityTransaction transaction = null;
-        try {
-            transaction = entityManager.getTransaction();
-            transaction.begin();
-            var customer = entityManager.find(Customer.class, customerId);
-            customer.setPhone(phone);
-            entityManager.merge(customer);
-            transaction.commit();
-            log.info("Customer is updated");
-        } catch (Exception e) {
-            if (transaction != null) {
-                transaction.rollback();
-            }
-            log.error("Customer is not updated. Exception is: ", e);
-        }
-    }
-
-    /**
-     * Delete customer from database
-     *
-     * @param customerId customer Id
-     */
-    public void delete(int customerId) {
-        EntityTransaction transaction = null;
-        try {
-            transaction = entityManager.getTransaction();
-            transaction.begin();
-            entityManager.remove(entityManager.find(Customer.class, customerId));
-            transaction.commit();
-        } catch (Exception e) {
-            if (transaction != null) {
-                transaction.rollback();
-            }
-            log.error("Customer is not deleted. Exception is: ", e);
-        }
-    }
-}
