--- assignment-impl/src/main/java/com/epam/rd/davydova/assignment/domain/service/ProductService.java
@@ -1,188 +1,0 @@
-package com.epam.rd.davydova.assignment.domain.service;
-
-import com.epam.rd.davydova.assignment.domain.model.Product;
-import com.epam.rd.davydova.assignment.domain.model.Supplier;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
-import javax.persistence.EntityManager;
-import javax.persistence.EntityTransaction;
-import java.math.BigDecimal;
-import java.util.List;
-import java.util.Optional;
-
-/**
- * This is a class for operations with Product entity and database
- */
-@Slf4j
-@Data
-@NoArgsConstructor
-@Service
-public class ProductService {
-    private EntityManager entityManager;
-
-    @Autowired
-    public ProductService(EntityManager entityManager) {
-        this.entityManager = entityManager;
-    }
-
-    /**
-     * Add product to database
-     *
-     * @param productName product name
-     * @param supplierId  supplier Id
-     * @param unitPrice price per unit
-     */
-    public void add(String productName, int supplierId, double unitPrice) {
-        EntityTransaction transaction = null;
-        try {
-            transaction = entityManager.getTransaction();
-            transaction.begin();
-            var product = new Product();
-            product.setProductName(productName);
-            product.setSupplier(entityManager.find(Supplier.class, supplierId));
-            product.setUnitPrice(BigDecimal.valueOf(unitPrice));
-            product.setDiscontinued(false);
-            entityManager.persist(product);
-            transaction.commit();
-            log.info("Product is added");
-        } catch (Exception e) {
-            if (transaction != null) {
-                transaction.rollback();
-            }
-            log.error("Product is not added. Exception is: ", e);
-        }
-    }
-
-    /**
-     * Find product by name
-     *
-     * @param productName product name
-     * @return Optional of product instance
-     */
-    public Optional<Product> findBy(String productName) {
-        EntityTransaction transaction = null;
-        Product product = null;
-        try {
-            transaction = entityManager.getTransaction();
-            transaction.begin();
-            product = (Product) entityManager
-                    .createNamedQuery(Product.FIND_PRODUCT_BY_NAME)
-                    .setParameter(1, productName)
-                    .getSingleResult();
-            transaction.commit();
-            var foundProduct = Optional.ofNullable(product);
-            if (foundProduct.isPresent()) {
-                log.info("Product is found");
-                return foundProduct;
-            }
-        } catch (Exception e) {
-            if (transaction != null) {
-                transaction.rollback();
-            }
-            log.error("Product is not found. Exception is: ", e);
-        }
-        return Optional.empty();
-    }
-
-    /**
-     * Find product by Id
-     *
-     * @param productId product Id
-     * @return Optional of product instance
-     */
-    public Optional<Product> findBy(int productId) {
-        EntityTransaction transaction = null;
-        try {
-            transaction = entityManager.getTransaction();
-            transaction.begin();
-            var product = entityManager.find(Product.class, productId);
-            transaction.commit();
-            var foundProduct = Optional.ofNullable(product);
-            if (foundProduct.isPresent()) {
-                log.info("Product is found");
-                return foundProduct;
-            }
-        } catch (Exception e) {
-            if (transaction != null) {
-                transaction.rollback();
-            }
-            log.error("Product is not found. Exception is: ", e);
-        }
-        return Optional.empty();
-    }
-
-    /**
-     * Find all products
-     *
-     * @return Optional of List of products
-     */
-    public Optional<List> findAll() {
-        EntityTransaction transaction = null;
-        try {
-            transaction = entityManager.getTransaction();
-            transaction.begin();
-            var productList = entityManager.createNamedQuery(Product.FIND_ALL_PRODUCTS).getResultList();
-            transaction.commit();
-            var foundProductList = Optional.ofNullable(productList);
-            if (foundProductList.isPresent()) {
-                log.info("List of products is found");
-                return foundProductList;
-            }
-        } catch (Exception e) {
-            if (transaction != null) {
-                transaction.rollback();
-            }
-            log.error("List of products is not found. Exception is: ", e);
-        }
-        return Optional.empty();
-    }
-
-    /**
-     * Update product status
-     *
-     * @param productId      product Id
-     * @param isDiscontinued product status
-     */
-    public void update(int productId, boolean isDiscontinued) {
-        EntityTransaction transaction = null;
-        try {
-            transaction = entityManager.getTransaction();
-            transaction.begin();
-            var product = entityManager.find(Product.class, productId);
-            product.setDiscontinued(isDiscontinued);
-            entityManager.merge(product);
-            transaction.commit();
-            log.info("Product is updated");
-        } catch (Exception e) {
-            if (transaction != null) {
-                transaction.rollback();
-            }
-            log.error("Product is not updated. Exception is: ", e);
-        }
-    }
-
-    /**
-     * Delete product from database
-     *
-     * @param productId product Id
-     */
-    public void delete(int productId) {
-        EntityTransaction transaction = null;
-        try {
-            transaction = entityManager.getTransaction();
-            transaction.begin();
-            entityManager.remove(entityManager.find(Product.class, productId));
-            transaction.commit();
-            log.info("Product is deleted");
-        } catch (Exception e) {
-            if (transaction != null) {
-                transaction.rollback();
-            }
-            log.error("Product is not deleted. Exception is: ", e);
-        }
-    }
-}
