+++ assignment-impl/src/main/java/com/epam/rd/davydova/assignment/service/stub/IService.java
@@ -1,0 +1,5 @@
+package com.epam.rd.davydova.assignment.service.stub;
+
+public interface IService {
+    String findAll(String languageTag);
+}
