+++ assignment-impl/src/main/java/com/epam/rd/davydova/assignment/service/stub/impl/ProductServiceStub.java
@@ -1,0 +1,33 @@
+package com.epam.rd.davydova.assignment.service.stub.impl;
+
+import com.epam.rd.davydova.assignment.domain.stub.ProductStub;
+import com.epam.rd.davydova.assignment.service.stub.IService;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.context.MessageSource;
+import org.springframework.stereotype.Service;
+
+import java.util.Locale;
+
+@Slf4j
+@Service
+@RequiredArgsConstructor
+public class ProductServiceStub implements IService {
+    private final ProductStub productStub;
+    private final MessageSource messageSource;
+
+    /**
+     * Find stubbed value of Product in RU or EN
+     *
+     * @param languageTag language tag (en-US or ru-RU)
+     * @return string representation of Product object
+     */
+    @Override
+    public String findAll(String languageTag) {
+        var result = messageSource.getMessage("message4",
+                new Object[]{productStub.getProductId(), productStub.getProductName(), productStub.getUnitPrice(), productStub.isDiscontinued()},
+                "message", Locale.forLanguageTag(languageTag));
+        log.info("{}", result);
+        return result;
+    }
+}
