+++ assignment-impl/src/main/java/com/epam/rd/davydova/assignment/service/stub/impl/SupplierServiceStub.java
@@ -1,0 +1,33 @@
+package com.epam.rd.davydova.assignment.service.stub.impl;
+
+import com.epam.rd.davydova.assignment.domain.stub.SupplierStub;
+import com.epam.rd.davydova.assignment.service.stub.IService;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.context.MessageSource;
+import org.springframework.stereotype.Service;
+
+import java.util.Locale;
+
+@Slf4j
+@Service
+@RequiredArgsConstructor
+public class SupplierServiceStub implements IService {
+    private final SupplierStub supplierStub;
+    private final MessageSource messageSource;
+
+    /**
+     * Find stubbed value of Supplier in RU or EN
+     *
+     * @param languageTag language tag (en-US or ru-RU)
+     * @return string representation of Supplier object
+     */
+    @Override
+    public String findAll(String languageTag) {
+        var result = messageSource.getMessage("message2",
+                new Object[]{supplierStub.getSupplierId(), supplierStub.getCompanyName(), supplierStub.getPhone()},
+                "message", Locale.forLanguageTag(languageTag));
+        log.info("{}", result);
+        return result;
+    }
+}
