--- assignment-impl/src/main/resources/properties/customer.properties
@@ -1,3 +1,0 @@
-customer_id=1
-customer_name=Massimo Warburton
-phone=94-53-80
