+++ assignment-impl/src/main/resources/properties/stub_customer.properties
@@ -1,0 +1,3 @@
+customer_id=1
+customer_name=Massimo Warburton
+customer_phone=94-53-80
