+++ assignment-impl/src/main/resources/properties/stub_product.properties
