+++ assignment-impl/src/main/resources/properties/stub_supplier.properties
@@ -1,0 +1,3 @@
+supplier_id=1
+company_name=SnailsCo
+supplier_phone=01-63-95
