--- assignment-impl/src/main/resources/properties/supplier.properties
@@ -1,3 +1,0 @@
-supplier_id=1
-company_name=SnailsCo
-phone=01-63-95
